refactor: split up mocking in the unit tests
