+++ src/tests/unit/mock_data/mock_blueprints/CarTest.blueprint.json
@@ -1,0 +1,118 @@
+{
+  "name": "CarTest",
+  "type": "system/SIMOS/Blueprint",
+  "attributes": [
+    {
+      "name": "name",
+      "type": "system/SIMOS/BlueprintAttribute",
+      "attributeType": "string",
+      "default": "CarTest"
+    },
+    {
+      "name": "type",
+      "type": "system/SIMOS/BlueprintAttribute",
+      "attributeType": "string",
+      "default": "test_data/complex/CarTest"
+    },
+    {
+      "name": "plateNumber",
+      "type": "system/SIMOS/BlueprintAttribute",
+      "attributeType": "string"
+    },
+    {
+      "name": "wheel",
+      "type": "system/SIMOS/BlueprintAttribute",
+      "attributeType": "test_data/complex/WheelTest"
+    },
+    {
+      "name": "wheels",
+      "type": "system/SIMOS/BlueprintAttribute",
+      "attributeType": "test_data/complex/WheelTest",
+      "dimensions": "*"
+    },
+    {
+      "name": "seats",
+      "type": "system/SIMOS/BlueprintAttribute",
+      "attributeType": "integer",
+      "default": 2
+    },
+    {
+      "name": "is_sedan",
+      "type": "system/SIMOS/BlueprintAttribute",
+      "attributeType": "boolean",
+      "default": true
+    },
+    {
+      "name": "floatValues",
+      "type": "system/SIMOS/BlueprintAttribute",
+      "attributeType": "number",
+      "dimensions": "*",
+      "default": [
+        2.1,
+        3.1,
+        4.2
+      ]
+    },
+    {
+      "name": "intValues",
+      "type": "system/SIMOS/BlueprintAttribute",
+      "attributeType": "integer",
+      "dimensions": "*",
+      "default": [
+        1,
+        5,
+        4,
+        2
+      ]
+    },
+    {
+      "name": "boolValues",
+      "type": "system/SIMOS/BlueprintAttribute",
+      "attributeType": "boolean",
+      "dimensions": "*",
+      "default": [
+        true,
+        false,
+        true
+      ]
+    },
+    {
+      "name": "stringValues",
+      "type": "system/SIMOS/BlueprintAttribute",
+      "attributeType": "string",
+      "dimensions": "*",
+      "default": [
+        "one",
+        "two",
+        "three"
+      ]
+    },
+    {
+      "name": "engine",
+      "type": "system/SIMOS/BlueprintAttribute",
+      "attributeType": "test_data/complex/EngineTest"
+    },
+    {
+      "name": "engine2",
+      "type": "system/SIMOS/BlueprintAttribute",
+      "optional": true,
+      "attributeType": "test_data/complex/EngineTest"
+    },
+    {
+      "name": "engine3",
+      "type": "system/SIMOS/BlueprintAttribute",
+      "optional": true,
+      "attributeType": "test_data/complex/EngineTest",
+      "default": {
+        "name": "default engine",
+        "type": "test_data/complex/EngineTest",
+        "fuelPump": {
+          "name": "fuelPump",
+          "type": "test_data/complex/FuelPumpTest",
+          "description": "A standard fuel pump"
+        },
+        "power": 9
+      }
+    }
+  ]
+}
