+++ src/tests/unit/mock_data/mock_blueprints/Customer.blueprint.json
@@ -1,0 +1,17 @@
+{
+  "name": "Customer",
+  "type": "system/SIMOS/Blueprint",
+  "attributes": [
+    {
+      "name": "name",
+      "type": "system/SIMOS/BlueprintAttribute",
+      "attributeType": "string"
+    },
+    {
+      "name": "car",
+      "type": "system/SIMOS/BlueprintAttribute",
+      "attributeType": "test_data/complex/RentalCar",
+      "contained": false
+    }
+  ]
+}
