+++ src/tests/unit/mock_data/mock_blueprints/EngineTest.blueprint.json
@@ -1,0 +1,38 @@
+{
+  "name": "EngineTest",
+  "type": "system/SIMOS/Blueprint",
+  "description": "This describes an engine",
+  "attributes": [
+    {
+      "name": "name",
+      "type": "system/SIMOS/BlueprintAttribute",
+      "attributeType": "string"
+    },
+    {
+      "name": "type",
+      "type": "system/SIMOS/BlueprintAttribute",
+      "attributeType": "string"
+    },
+    {
+      "name": "description",
+      "type": "system/SIMOS/BlueprintAttribute",
+      "attributeType": "string"
+    },
+    {
+      "name": "power",
+      "type": "system/SIMOS/BlueprintAttribute",
+      "attributeType": "integer",
+      "default": 120
+    },
+    {
+      "name": "fuelPump",
+      "type": "system/SIMOS/BlueprintAttribute",
+      "attributeType": "test_data/complex/FuelPumpTest",
+      "default": {
+        "name": "fuelPump",
+        "type": "test_data/complex/FuelPumpTest",
+        "description": "A standard fuel pump"
+      }
+    }
+  ]
+}
