+++ src/tests/unit/mock_data/mock_blueprints/ExtendedBlueprint.blueprint.json
@@ -1,0 +1,36 @@
+{
+  "name": "ExtendedBlueprint",
+  "type": "dmss://system/SIMOS/Blueprint",
+  "description": "This Blueprint extends blueprint 2",
+  "extends": [
+    "basic_blueprint"
+  ],
+  "attributes": [
+    {
+      "name": "another_value",
+      "type": "dmss://system/SIMOS/BlueprintAttribute",
+      "attributeType": "string"
+    }
+  ],
+  "storageRecipes": [
+    {
+      "name": "default",
+      "type": "dmss://system/SIMOS/StorageRecipe",
+      "description": "",
+      "storageAffinity": "blob",
+      "attributes": []
+    }
+  ],
+  "uiRecipes": [
+    {
+      "name": "default",
+      "description": "",
+      "plugin": "edit2"
+    },
+    {
+      "name": "aSpecialView",
+      "description": "",
+      "plugin": "specialView"
+    }
+  ]
+}
