+++ src/tests/unit/mock_data/mock_blueprints/FuelPumpTest.blueprint.json
@@ -1,0 +1,23 @@
+{
+  "name": "FuelPumpTest",
+  "type": "system/SIMOS/Blueprint",
+  "description": "This describes a fuel pump",
+  "attributes": [
+    {
+      "name": "name",
+      "type": "system/SIMOS/BlueprintAttribute",
+      "attributeType": "string"
+    },
+    {
+      "name": "description",
+      "type": "system/SIMOS/BlueprintAttribute",
+      "attributeType": "string",
+      "default": "A standard fuel pump"
+    },
+    {
+      "name": "type",
+      "type": "system/SIMOS/BlueprintAttribute",
+      "attributeType": "string"
+    }
+  ]
+}
