+++ src/tests/unit/mock_data/mock_blueprints/Parent.blueprint.json
@@ -1,0 +1,15 @@
+{
+  "name": "Parent",
+  "type": "dmss://system/SIMOS/Blueprint",
+  "description": "",
+  "extends": [
+    "dmss://system/SIMOS/NamedEntity"
+  ],
+  "attributes": [
+    {
+      "name": "SomeChild",
+      "type": "dmss://system/SIMOS/BlueprintAttribute",
+      "attributeType": "BaseChild"
+    }
+  ]
+}
