+++ src/tests/unit/mock_data/mock_blueprints/WheelTest.blueprint.json
@@ -1,0 +1,23 @@
+{
+  "name": "WheelTest",
+  "type": "system/SIMOS/Blueprint",
+  "attributes": [
+    {
+      "name": "name",
+      "type": "system/SIMOS/BlueprintAttribute",
+      "attributeType": "string",
+      "default": "Wheel"
+    },
+    {
+      "name": "type",
+      "type": "system/SIMOS/BlueprintAttribute",
+      "attributeType": "string"
+    },
+    {
+      "name": "power",
+      "type": "system/SIMOS/BlueprintAttribute",
+      "attributeType": "number",
+      "default": 0.0
+    }
+  ]
+}
