+++ src/tests/unit/mock_data/mock_blueprints/all_contained_cases_blueprint.blueprint.json
@@ -1,0 +1,26 @@
+{
+  "name": "Blueprint 1",
+  "type": "dmss://system/SIMOS/Blueprint",
+  "description": "First blueprint",
+  "extends": [
+    "dmss://system/SIMOS/NamedEntity"
+  ],
+  "attributes": [
+    {
+      "name": "nested",
+      "type": "dmss://system/SIMOS/BlueprintAttribute",
+      "attributeType": "basic_blueprint"
+    },
+    {
+      "name": "reference",
+      "type": "dmss://system/SIMOS/BlueprintAttribute",
+      "attributeType": "basic_blueprint"
+    },
+    {
+      "name": "references",
+      "type": "dmss://system/SIMOS/BlueprintAttribute",
+      "attributeType": "basic_blueprint",
+      "dimensions": "*"
+    }
+  ]
+}
