+++ src/tests/unit/mock_data/mock_blueprints/basic_blueprint.blueprint.json
@@ -1,0 +1,17 @@
+{
+  "name": "Blueprint 2",
+  "type": "dmss://system/SIMOS/Blueprint",
+  "description": "Second blueprint",
+  "extends": [
+    "dmss://system/SIMOS/NamedEntity"
+  ],
+  "attributes": [],
+  "storageRecipes": [],
+  "uiRecipes": [
+    {
+      "name": "default",
+      "description": "",
+      "plugin": "edit1"
+    }
+  ]
+}
