+++ src/tests/unit/mock_data/mock_blueprints/blob.blueprint.json
@@ -1,0 +1,24 @@
+{
+  "name": "blob",
+  "type": "dmss://system/SIMOS/Blueprint",
+  "description": "Test for a blueprint with storageAffinity in storageRecipe",
+  "extends": [
+    "dmss://system/SIMOS/NamedEntity"
+  ],
+  "attributes": [
+    {
+      "name": "someData",
+      "type": "dmss://system/SIMOS/BlueprintAttribute",
+      "attributeType": "string"
+    }
+  ],
+  "storageRecipes": [
+    {
+      "name": "default",
+      "type": "dmss://system/SIMOS/StorageRecipe",
+      "description": "",
+      "storageAffinity": "blob",
+      "attributes": []
+    }
+  ]
+}
