+++ src/tests/unit/mock_data/mock_blueprints/blueprint_with_nested_optional_attr.blueprint.json
@@ -1,0 +1,15 @@
+{
+  "name": "blueprint_with_nested_optional_attr",
+  "type": "dmss://system/SIMOS/Blueprint",
+  "description": "",
+  "extends": [
+    "dmss://system/SIMOS/NamedEntity"
+  ],
+  "attributes": [
+    {
+      "name": "nested_with_optional",
+      "type": "dmss://system/SIMOS/BlueprintAttribute",
+      "attributeType": "blueprint_with_optional_attr"
+    }
+  ]
+}
