+++ src/tests/unit/mock_data/mock_blueprints/blueprint_with_optional_attr.blueprint.json
@@ -1,0 +1,17 @@
+{
+  "name": "Blueprint_with_optional_attr",
+  "type": "dmss://system/SIMOS/Blueprint",
+  "description": "",
+  "extends": [
+    "dmss://system/SIMOS/NamedEntity"
+  ],
+  "attributes": [
+    {
+      "name": "im_optional",
+      "type": "dmss://system/SIMOS/BlueprintAttribute",
+      "attributeType": "basic_blueprint",
+      "optional": true,
+      "contained": true
+    }
+  ]
+}
