+++ src/tests/unit/mock_data/mock_blueprints/blueprint_with_second_level_nested_uncontained_attribute.blueprint.json
@@ -1,0 +1,15 @@
+{
+  "name": "blueprint_with_second_level_nested_uncontained_attribute",
+  "type": "dmss://system/SIMOS/Blueprint",
+  "description": "",
+  "extends": [
+    "dmss://system/SIMOS/NamedEntity"
+  ],
+  "attributes": [
+    {
+      "name": "i_have_a_uncontained_attribute",
+      "type": "dmss://system/SIMOS/BlueprintAttribute",
+      "attributeType": "uncontained_blueprint"
+    }
+  ]
+}
