+++ src/tests/unit/mock_data/mock_blueprints/two_contained_deep_attributes.blueprint.json
@@ -1,0 +1,20 @@
+{
+  "name": "two_contained_deep_attributes",
+  "type": "dmss://system/SIMOS/Blueprint",
+  "description": "Two contained deeply nested attributes",
+  "extends": [
+    "dmss://system/SIMOS/NamedEntity"
+  ],
+  "attributes": [
+    {
+      "name": "a",
+      "type": "dmss://system/SIMOS/BlueprintAttribute",
+      "attributeType": "all_contained_cases_blueprint"
+    },
+    {
+      "name": "b",
+      "type": "dmss://system/SIMOS/BlueprintAttribute",
+      "attributeType": "all_contained_cases_blueprint"
+    }
+  ]
+}
