+++ src/tests/unit/mock_data/mock_blueprints/uncontained_blueprint.blueprint.json
@@ -1,0 +1,16 @@
+{
+  "name": "uncontained_blueprint",
+  "type": "dmss://system/SIMOS/Blueprint",
+  "description": "uncontained_blueprint",
+  "extends": [
+    "dmss://system/SIMOS/NamedEntity"
+  ],
+  "attributes": [
+    {
+      "name": "uncontained_in_every_way",
+      "type": "dmss://system/SIMOS/BlueprintAttribute",
+      "attributeType": "basic_blueprint",
+      "contained": false
+    }
+  ]
+}
