+++ src/tests/unit/test_create_default_array.py
@@ -8,7 +8,7 @@ from domain_classes.dimension import Dimension
 from enums import SIMOS
 from services.document_service import DocumentService
 from storage.repositories.file import LocalFileRepository
-from tests.unit.mock_utils import mock_storage_recipe_provider
+from tests.unit.mock_data.mock_recipe_provider import mock_storage_recipe_provider
 
 basic_blueprint = {
     "type": SIMOS.BLUEPRINT.value,
