+++ src/tests/unit/test_get_extended_blueprint.py
@@ -5,7 +5,7 @@ from common.address import Address
 from common.utils.data_structure.compare import get_and_print_diff
 from domain_classes.blueprint import Blueprint
 from domain_classes.tree_node import Node
-from tests.unit.mock_utils import get_mock_document_service
+from tests.unit.mock_data.mock_document_service import get_mock_document_service
 
 
 class GetExtendedBlueprintTestCase(unittest.TestCase):
