+++ src/tests/unit/test_validate_entity.py
@@ -3,10 +3,8 @@ import unittest
 from common.exceptions import ValidationException
 from common.tree_node_serializer import tree_node_from_dict, tree_node_to_dict
 from common.utils.validators import validate_entity, validate_entity_against_self
-from tests.unit.mock_utils import (
-    get_mock_document_service,
-    mock_storage_recipe_provider,
-)
+from tests.unit.mock_data.mock_document_service import get_mock_document_service
+from tests.unit.mock_data.mock_recipe_provider import mock_storage_recipe_provider
 
 
 class ValidateEntityTestCase(unittest.TestCase):
